Restructure the project and reconfigure setup.py
Project filetree is completely restructured, package `create_flask_app`
is now inside `src` folder.

File `setup.py` is now configured and works well. It recognizes
previously mentioned package and in conjunction with `MANIFEST.in`
copies all template folders to its designated directory when installing
the package.
