+++ .flake8
@@ -1,0 +1,9 @@
+[flake8]
+ignore = E501
+exclude =
+    .git,
+    __pycache__,
+    old,
+    build,
+    dist
+max-complexity = 10
