+++ MANIFEST.in
@@ -1,0 +1,4 @@
+recursive-include src/create_flask_app/api *
+recursive-include src/create_flask_app/default *
+recursive-include src/create_flask_app/helloworld *
+recursive-include src/create_flask_app/spa *
