+++ setup.py
@@ -1,11 +1,13 @@
+import glob
 import setuptools
+from os.path import basename, splitext
 
 with open("README.md", "r", encoding="utf8") as fh:
     long_description = fh.read()
 
 setuptools.setup(
     name="createflaskapp",
-    version="1.0",
+    version="1.2",
     license="MIT",
     author="Ivan Sakal",
     author_email="sakalivan4@gmail.com",
@@ -21,13 +23,17 @@ setuptools.setup(
         "Operating System :: OS Independent",
     ],
     python_requires=">=3.6",
-    pymodules=["create-flask-app"],
+    package_dir={"": "src"},
+    packages=setuptools.find_packages("src"),
+    py_modules=[splitext(basename(path))[0] for path in glob.glob('src/*.py')],
+    include_package_data=True,
+    # package_data={"": [".flake8"]},
     install_requires=[
         "Click>=7",
         "animation>=0.0.6",
         "termcolor>=1.1.0"
     ],
     entry_points={
-        "console_scripts": ["create-flask-app=create_flask_app:create_project"]
+        "console_scripts": ["create-flask-app=create_flask_app.cli:create_project"]
     }
 )
