+++ src/create_flask_app/__main__.py
@@ -1,4 +1,4 @@
-from create_flask_app import create_project
+from create_flask_app.cli import create_project
 
 if __name__ == '__main__':
     create_project()
