+++ src/create_flask_app/spa/app/config.py
