+++ src/create_flask_app/spa/app/static/css/main.css
