+++ src/create_flask_app/spa/app/static/js/button.js
