+++ src/create_flask_app/spa/app/templates/base.html
