+++ src/create_flask_app/spa/app/templates/home.html
