+++ src/create_flask_app/spa/requirements.txt
